+++ dipy/utils/meson.build
@@ -27,7 +27,6 @@ python_sources = [
   '__init__.py',
   'arrfuncs.py',
   'convert.py',
-  'defaults.py',
   'deprecator.py',
   'multiproc.py',
   'optpkg.py',
