+++ dipy/utils/tests/meson.build
@@ -20,7 +20,6 @@ python_sources = [
   '__init__.py',
   'test_arrfuncs.py',
   'test_convert.py',
-  'test_defaults.py',
   'test_deprecator.py',
   'test_multiproc.py',
   'test_omp.py',
